uses dbconfig for Debian based distributions
dbconfig-common is a standard mechanism for Debian based distributions,
to assists applications that require a database.
The user still have to choice to configure the database manually,
if not, dbconfig detects want database backends are installed,
and the user gets ask, want connection parameter he wants to use.
If dbconfig is configured correctly, database schema updated are done automatically on package updates.
With
export DEBIAN_FRONTEND=noninteractive
the user don't get ask, and dbconfig configures the database backend as good as he can.